Done with gate authorizations

+++ app/Http/Controllers/QuestionsController.php
@@ -6,6 +6,8 @@ use App\Models\Question;
 use Illuminate\Http\Request;
 use App\Http\Requests\AskQuestionRequest;
 
+use Illuminate\Support\Facades\Gate;
+
 class QuestionsController extends Controller
 {
     /**
@@ -73,7 +75,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+        if (Gate::allows('update-question', $question)) {
+            return view('questions.edit', compact('question'));
+        } else {
+            abort(403, "You are not allowed to edit");
+        }
     }
 
     /**
@@ -85,6 +91,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403, "You are not allowed to update");
+        } 
+
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', 'Your question has been updated!');
@@ -98,6 +108,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question', $question)) {
+            abort(403, "You are not allowed to delete");
+        } 
+
         $question->delete();
 
         return redirect('/questions')->with('success', "Your question has been deleted!");
